+++ gulp/config.js
@@ -1,4 +1,4 @@
-export const pkgJson = require('./package.json');
+export const pkgJson = require('../package.json');
 
 export const mainDirectories = {
 	dev: './server/',
@@ -11,7 +11,9 @@ export const settings = {
 		styles: ['./src/assets/scss/**/*.scss'],
 		stylesEntryPoint: './src/assets/scss/index.scss',
 		scripts: ['./src/app/**/*.js'],
+		appTemplates: ['./src/app/**/*.html'],
 		images: ['./src/assets/img/**/*.{png,jpg,gif,svg}'],
+		fonts: ['./src/assets/fonts/**/*'],
 		externalCss: pkgJson.bootstrapKickstart.bundleCSS,
 		externalJs: pkgJson.bootstrapKickstart.bundleExternalJS,
 		staticFiles: pkgJson.bootstrapKickstart.includeStaticFiles
@@ -20,26 +22,18 @@ export const settings = {
 		dev: {
 			markup: `${mainDirectories.dev}`,
 			styles: `${mainDirectories.dev}assets/css/`,
-			scripts: `${mainDirectories.dev}app/`,
+			app: `${mainDirectories.dev}app/`,
+			fonts: `${mainDirectories.dev}assets/fonts/`,
 			images: `${mainDirectories.dev}assets/img/`,
 			libs: `${mainDirectories.dev}libs/`
 		},
 		prod: {
 			markup: `${mainDirectories.dist}`,
 			styles: `${mainDirectories.dist}assets/css/`,
-			scripts: `${mainDirectories.dist}app/`,
+			app: `${mainDirectories.dist}app/`,
+			fonts: `${mainDirectories.dist}assets/fonts/`,
 			images: `${mainDirectories.dist}assets/img/`,
 			libs: `${mainDirectories.dist}libs/`
 		}
-	},
-	autoPrefix: {
-		browsers: [
-			'> 1%',
-			'last 3 version',
-			'ie 8',
-			'ie 9',
-			'Firefox ESR',
-			'Opera 12.1'
-		]
 	}
 };